function-source/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-aiplatform==1.*
vertexai==1.*
google-cloud-bigquery
requests

